Fix exposure slider import case conflict
